requirements.txt
aiogram>=3.0.0
qrcode>=7.4.2
Pillow>=10.0.0
opencv-python-headless>=4.8.0
python-dotenv>=1.0.0
pystyle